requirements.txt
annotated-types==0.7.0
anyio==4.8.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpeg==1.4
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==1.59.7
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tomli==2.2.1
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
